+++ Changes.md
@@ -23,6 +23,7 @@ API
   - Deprecated `allowMultipleSelection` constructor argument. Use `selectionMode` argument instead.
   - Added modes to `selectionMode` allowing single and multiple cell selections.
   - `setSelection` and `getSelection` now accept and return different types depending on the `selectionMode`. For `Row` and `Rows` modes, a single `PathMatcher` object represents the selection for all columns, maintaining backwards compatiblity. For `Cell` and `Cells` modes, a list of `PatchMatcher` objects, one per column, represents the selection.
+- PlugValueWidget : Added new exception types `MultipleWidgetCreatorsError` and `MultiplePlugTypesError`.
 
 0.61.13.1 (relative to 0.61.13.0)
 =========